+++ continuity.service.cobra/src/main/java/org/continuity/cobra/managers/ElasticsearchSessionManager.java
@@ -53,7 +53,7 @@ public class ElasticsearchSessionManager extends ElasticsearchScrollingManager<S
 
 	private boolean updateScriptInitialized = false;
 
-	public ElasticsearchSessionManager(String host, ObjectMapper mapper, long bulkTimeoutSeconds) throws IOException {
+	public ElasticsearchSessionManager(String host, ObjectMapper mapper, int bulkTimeoutSeconds) throws IOException {
 		super(host, "session", bulkTimeoutSeconds);
 
 		this.mapper = mapper;
