+++ continuity.service.cobra/src/main/java/org/continuity/cobra/managers/ElasticsearchTraceManager.java
@@ -35,7 +35,7 @@ public class ElasticsearchTraceManager extends ElasticsearchScrollingManager<Tra
 
 	private final ObjectMapper mapper;
 
-	public ElasticsearchTraceManager(String host, ObjectMapper mapper, long bulkTimeoutSeconds) throws IOException {
+	public ElasticsearchTraceManager(String host, ObjectMapper mapper, int bulkTimeoutSeconds) throws IOException {
 		super(host, "trace", bulkTimeoutSeconds);
 		this.mapper = mapper;
 	}
